requirements.txt
adal==1.2.7
ansiwrap==0.8.4
argcomplete==1.12.3
attrs==21.4.0
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azureml-core==1.40.0.post2
azureml-dataprep==3.1.1
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.5.2
azureml-mlflow==1.40.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
cryptography==36.0.2
cycler==0.11.0
databricks-cli==0.16.6
distro==1.7.0
docker==5.0.3
dotnetcore2==2.1.23
entrypoints==0.4
fastjsonschema==2.15.3
fonttools==4.33.2
gitdb==4.0.9
GitPython==3.1.27
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
isodate==0.6.1
jeepney==0.8.0
jmespath==0.10.0
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.2
knack==0.9.0
lightgbm==3.3.2
matplotlib==3.5.1
mlflow-skinny==1.25.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
mypy-extensions==0.4.3
nbclient==0.6.0
nbformat==5.3.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
papermill==2.3.4
paramiko==2.10.3
pathspec==0.9.0
Pillow==9.1.0
pkginfo==1.8.2
platformdirs==2.5.2
portalocker==2.4.0
protobuf==3.20.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==227; sys_platform=="darwin"
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==3.3.2
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
tenacity==8.0.1
textwrap3==0.9.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.7
websocket-client==1.3.2
zipp==3.8.0
